set the _id attribute manually for inserted user and articles

+++ routes/posts_routes.py
@@ -1,5 +1,5 @@
 from flask import jsonify, request
-from . import bp, storage, swag_from  # Imported from __init__.py
+from routes import bp, storage, swag_from, redis_client, Auth  # Imported from __init__.py
 from datetime import datetime
 
 
@@ -9,6 +9,11 @@ def get_current_date():
     """
     current_date = datetime.now()
     return current_date.strftime('%Y-%m-%d')
+
+
+def get_author_id():
+    token = redis_client.get(request.headers.get('Authorization'))
+    return token
 
 
 # GET /articles - returns a list of all saved posts
@@ -23,43 +28,45 @@ def get_all_posts():
         description: Returns a list of all posts.
     """
     posts = storage.find_posts_by_author()
-    return jsonify(posts)
+    for post in posts:
+        post['_id'] = str(post['_id'])
+    return jsonify(posts), 200
 
 
-# returns a list of posts with the given author_id
-@bp.route('/articles/<string:author_id>', methods=['GET'])
+@bp.route('/articles/<string:article_id>', methods=['GET'])
 @swag_from(methods=['GET'])
-def get_posts_by_author(author_id):
+def get_posts_by_id(article_id):
     """
-    Get posts by author.
+    Get a post using its article_id.
     ---
     parameters:
-      - name: author_id
+      - name: article_id
         in: path
         type: string
         required: true
-        description: The author's ID.
+        description: article's ID.
     responses:
       200:
-        description: Returns a list of posts by the specified author.
+        description: Returns a posts by the specified id.
     """
-    posts = storage.find_posts_by_author(author_id)
-    return jsonify(posts)
+    post = storage.find_post_by_id(article_id)
+    post['_id'] = str(post['_id'])
+    print(post)
+    return jsonify(post), 200
 
 
-# POST /articles/<str:author_id> - inserts the article to the db
-@bp.route('/articles/<string:author_id>', methods=['POST'])
+@bp.route('/articles', methods=['POST'])
 @swag_from(methods=['POST'])
-def insert_article(author_id):
+def insert_article():
     """
     Insert a new article.
     ---
     parameters:
-      - name: author_id
-        in: path
-        type: string
+      - name: Authorization
+        in: header
+        description: The authorization token.
         required: true
-        description: The author's ID.
+        type: string
       - name: title
         in: formData
         type: string
@@ -74,29 +81,30 @@ def insert_article(author_id):
       201:
         description: id of the newly created article.
     """
+    # author_id = redis_client.get(request.headers.get('Authorization'))
     article_data = {
+        "_id": Auth.get_token(),
         "title": request.form.get("title"),
         "content": request.form.get("content"),
-        "author_id": author_id,
+        "author_id": get_author_id(),
         "created_at": get_current_date()
     }
     result = storage.insert_post(article_data)
     return jsonify(str(result.inserted_id)), 201
 
 
-# DELETE /articles/<str:author_id> - delete an article using article_id
-@bp.route('/articles/<string:author_id>/<string:article_id>', methods=['DELETE'])
+@bp.route('/articles/<string:article_id>', methods=['DELETE'])
 @swag_from(methods=['DELETE'])
-def delete_article(author_id, article_id):
+def delete_article(article_id):
     """
     Delete an article by its ID.
     ---
     parameters:
-      - name: author_id
-        in: path
-        type: string
+      - name: Authorization
+        in: header
+        description: The authorization token.
         required: true
-        description: The ID of the owner of the article.
+        type: string
       - name: article_id
         in: path
         type: string
@@ -106,26 +114,32 @@ def delete_article(author_id, article_id):
       204:
         description: The article was successfully deleted.
     """
-    result = storage.delete_post(article_id)
-    if result.deleted_count == 1:
-        return "Article successfully deleted", 204
-    else:
-        return jsonify({"error": "Article not found"}), 404
+    auth = False
+    for post in storage.find_posts_by_author(get_author_id()):
+        if str(post['_id']) == article_id:
+            auth = True
+
+    if auth:
+        result = storage.delete_post(article_id)
+        if result.deleted_count == 1:
+            return "Article successfully deleted", 204
+        else:
+            return jsonify({"error": "Article not found"}), 404
+    return 'Forbidden request', 403
 
 
-# PUT /articles/<str:author_id> - update the content of an article
-@bp.route('/articles/<string:author_id>/<string:article_id>', methods=['PUT'])
+@bp.route('/articles/<string:article_id>', methods=['PUT'])
 @swag_from(methods=['PUT'])
-def update_article_content(author_id, article_id):
+def update_article_content(article_id):
     """
     Update the content of an article by its ID.
     ---
     parameters:
-      - name: author_id
-        in: path
-        type: string
+      - name: Authorization
+        in: header
+        description: The authorization token.
         required: true
-        description: The ID of the author of the article.
+        type: string
       - name: article_id
         in: path
         type: string
@@ -142,11 +156,13 @@ def update_article_content(author_id, article_id):
       404:
         description: article not found
     """
-    updated_data = {"content": request.form.get("content"),
-                    "updated_at": get_current_date()}
-    # id = request.form.get("article_id")
-    result = storage.update_post(article_id, updated_data)
-    if result.modified_count == 1:
-        return jsonify({"message": "Article updated successfully"})
-    else:
-        return jsonify({"error": "Article not found"}), 404
+    if article_id in storage.find_posts_by_author(get_author_id()):
+        updated_data = {"content": request.form.get("content"),
+                        "updated_at": get_current_date()}
+        # id = request.form.get("article_id")
+        result = storage.update_post(article_id, updated_data)
+        if result.modified_count == 1:
+            return jsonify({"message": "Article updated successfully"})
+        else:
+            return jsonify({"error": "Article not found"}), 404
+    return "Forbidden request", 403
